Made Adjustments with Future Scenes
I added the hidden remarks for future scenes.

I fixed spacing and commenting on what does what. For some reason, i am getting a complie error and I cannot determine the cause.

I also added the Story.txt file of what is in the code for a more writing format.